Add permission checks. Minor fixes

+++ books_db/api/views.py
@@ -3,6 +3,8 @@ from flask_jwt import jwt_required, current_identity
 
 from api import serializers
 from api import validators
+
+from api.permissions import check_object_permission, check_self_permission
 from books.models import Book, Author
 from users.models import User
 
@@ -43,6 +45,7 @@ def update_book(book_id):
     if not request.json:
         abort(make_response(jsonify({'400': 'Bad request, not a JSON'}), 400))
     book = Book.get_or_404(book_id)
+    check_object_permission(current_identity, book)
     book.update(**validators.book_validator(request.json, True, book))
     return jsonify(serializers.book_serializer(book)), 201
 
@@ -51,6 +54,7 @@ def update_book(book_id):
 @jwt_required()
 def delete_book(book_id):
     book = Book.get_or_404(book_id)
+    check_object_permission(current_identity, book)
     book.delete()
     return make_response(
         jsonify(message=f"Book '{book.title}' has been deleted"), 204
@@ -60,12 +64,13 @@ def delete_book(book_id):
 @api.route('/api/authors', methods=('GET',))
 def get_authors():
     authors = Author.query.all()
-    serialized_data = [serializers.author_serializer(obj) for obj in authors]
+    serialized_data = [serializers.author_serializer(obj)
+                       for obj in authors]
     return jsonify(serialized_data)
 
 
 @api.route('/api/authors/<int:author_id>', methods=('GET',))
-def authors(author_id):
+def get_author(author_id):
     author = Author.get_or_404(author_id)
     return jsonify(serializers.author_serializer(author))
 
@@ -87,6 +92,7 @@ def update_author(author_id):
     if not request.json:
         abort(make_response(jsonify({'400': 'Bad request, not a JSON'}), 400))
     author = Author.get_or_404(author_id)
+    check_object_permission(current_identity, author)
     author.update(**validators.author_validator(request.json, True, author))
     return jsonify(serializers.author_serializer(author)), 201
 
@@ -95,6 +101,7 @@ def update_author(author_id):
 @jwt_required()
 def delete_author(author_id):
     author = Author.get_or_404(author_id)
+    check_object_permission(current_identity, author)
     author.delete()
     return make_response(
         jsonify(message=f"Author '{author.fio}' has been deleted"), 204
@@ -132,6 +139,7 @@ def update_user(user_id):
     if not request.json:
         abort(make_response(jsonify({'400': 'Bad request, not a JSON'}), 400))
     user = User.get_or_404(user_id)
+    check_self_permission(current_identity, user)
     user.update(**validators.user_validator(request.json, True, user))
     return jsonify(serializers.user_serializer(user)), 201
 
@@ -140,6 +148,7 @@ def update_user(user_id):
 @jwt_required()
 def delete_user(user_id):
     user = User.get_or_404(user_id)
+    check_self_permission(current_identity, user)
     user.delete()
     return make_response(
         jsonify(message=f"Author '{user.name}' has been deleted"), 204
